layerx-python-backend/requirements.txt
numpy
flask
flask-cors
tensorflow
# bson
pymongo
boto3
imgaug
# skimage
scikit-image
# opencv-python
Pillow
